+++ src/main/java/com/checkmarx/flow/custom/ADOIssueTracker.java
@@ -6,6 +6,7 @@ import com.checkmarx.flow.dto.Issue;
 import com.checkmarx.flow.dto.ScanRequest;
 import com.checkmarx.flow.dto.azure.CreateWorkItemAttr;
 import com.checkmarx.flow.exception.MachinaException;
+import com.checkmarx.flow.utils.HTMLHelper;
 import com.checkmarx.flow.utils.ScanUtils;
 import com.checkmarx.sdk.config.Constants;
 import com.checkmarx.sdk.dto.ScanResults;
@@ -234,7 +235,7 @@ public class ADOIssueTracker implements IssueTracker {
         CreateWorkItemAttr description = new CreateWorkItemAttr();
         description.setOp("add");
         description.setPath(Constants.ADO_FIELD.concat(FIELD_PREFIX.concat(issueBody)));
-        description.setValue(ScanUtils.getHTMLBody(resultIssue, request, flowProperties));
+        description.setValue(HTMLHelper.getHTMLBody(resultIssue, request, flowProperties));
         CreateWorkItemAttr tagsBlock = new CreateWorkItemAttr();
         tagsBlock.setOp("add");
         tagsBlock.setPath(Constants.ADO_FIELD.concat(TAGS_FIELD));
@@ -373,7 +374,7 @@ public class ADOIssueTracker implements IssueTracker {
         CreateWorkItemAttr description = new CreateWorkItemAttr();
         description.setOp("add");
         description.setPath(Constants.ADO_FIELD.concat(FIELD_PREFIX.concat(issueBody)));
-        description.setValue(ScanUtils.getHTMLBody(resultIssue, request, flowProperties));
+        description.setValue(HTMLHelper.getHTMLBody(resultIssue, request, flowProperties));
 
         List<CreateWorkItemAttr> body = new ArrayList<>(Arrays.asList(state, description));
