+++ src/main/java/com/checkmarx/flow/service/ASTScanner.java
@@ -6,21 +6,24 @@ import com.checkmarx.sdk.config.AstProperties;
 
 import com.checkmarx.sdk.dto.ScanResults;
 
+import com.checkmarx.sdk.dto.ast.ASTResults;
 import com.checkmarx.sdk.dto.ast.ASTResultsWrapper;
 
 
 import com.cx.restclient.AstClientImpl;
 
+import com.cx.restclient.ast.dto.sast.AstSastResults;
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
 
 
 @Service
 
 @Slf4j
-public class ASTScanner extends AbstractASTScanner  {
+public class ASTScanner extends AbstractASTScanner {
 
     public ASTScanner(AstClientImpl astClient, FlowProperties flowProperties) {
         super(astClient, flowProperties, AstProperties.CONFIG_PREFIX);
@@ -35,12 +38,13 @@ public class ASTScanner extends AbstractASTScanner  {
 
     @Override
     protected String getScanId(ASTResultsWrapper internalResults) {
-        if (internalResults.getAstResults() != null && internalResults.getAstResults().getResults() !=null){
-            return internalResults.getAstResults().getResults().getScanId();
-        }else{
-            return "";
-        }
+
+        return Optional.ofNullable(internalResults.getAstResults())
+                .map(ASTResults::getResults)
+                .map(AstSastResults::getScanId)
+                .orElse("");
+
     }
-    
+
 
 }
