+++ src/main/java/com/checkmarx/flow/service/FlowService.java
@@ -1,6 +1,7 @@
 package com.checkmarx.flow.service;
 
 import com.checkmarx.flow.dto.ScanRequest;
+import com.checkmarx.flow.exception.MachinaRuntimeException;
 import com.checkmarx.sdk.dto.ScanResults;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -34,7 +35,16 @@ public class FlowService {
         String effectiveProjectName = projectNameGenerator.determineProjectName(scanRequest);
         scanRequest.setProject(effectiveProjectName);
         List<VulnerabilityScanner> enabledScanners = getEnabledScanners(scanRequest);
+        validateEnabledScanners(enabledScanners);
         runScanRequest(scanRequest, enabledScanners);
+    }
+
+    private void validateEnabledScanners(List<VulnerabilityScanner> enabledScanners) {
+        boolean isSastAndASTScannersFound = enabledScanners.stream().anyMatch(scanner -> scanner instanceof ASTScanner)
+                && enabledScanners.stream().anyMatch(scanner -> scanner instanceof SastScanner);
+        if (isSastAndASTScannersFound) {
+            throw new MachinaRuntimeException("Both SAST & AST-SAST scanners cannot be set together");
+        }
     }
 
     private void runScanRequest(ScanRequest scanRequest, List<VulnerabilityScanner> scanners) {
