+++ src/test/resources/cucumber/data/cxflow-reference-reports/1-finding.json
@@ -55,5 +55,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
