+++ src/test/resources/cucumber/data/cxflow-reference-reports/11-findings.json
@@ -137,5 +137,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
