+++ src/test/resources/cucumber/data/cxflow-reference-reports/2-findings-different-severity.json
@@ -76,5 +76,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
