+++ src/test/resources/cucumber/data/cxflow-reference-reports/2-findings-different-vuln-type-same-file.json
@@ -97,5 +97,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
