+++ src/test/resources/cucumber/data/cxflow-reference-reports/2-findings-same-severity.json
@@ -75,5 +75,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
