+++ src/test/resources/cucumber/data/cxflow-reference-reports/2-findings-same-vuln-type-different-files.json
@@ -96,5 +96,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
