+++ src/test/resources/cucumber/data/cxflow-reference-reports/2-findings-same-vuln-type-same-file.json
@@ -75,5 +75,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
