+++ src/test/resources/cucumber/data/cxflow-reference-reports/3-findings-different-severity-medium-high-critical.json
@@ -55,5 +55,6 @@
       },
       "allFalsePositive": false
     }
-  ]
+  ],
+  "sastRestuls": false
 }
