+++ src/test/resources/cucumber/data/cxflow-reference-reports/empty-results.json
@@ -11,5 +11,6 @@
     "scanId": "1000024",
     "scanStartDate": "Wednesday, January 15, 2020 1:31:13 PM"
   },
-  "xissues": []
+  "xissues": [],
+  "sastRestuls": false
 }
